+++ src/System.Diagnostics.Debug/tests/DebugTests.cs
@@ -11,38 +11,219 @@ namespace System.Diagnostics.Tests
     [Collection("System.Diagnostics.Debug")]
     public class DebugTests
     {
+        // When true, tests Debug after it has wired up with Trace Listeners
+        private static readonly bool _testDebugWithTraceListeners = true;
+
         [Fact]
-        public void Write_Indents()
+        public void Debug_Write_Indents()
         {
             // This test when run alone verifies Debug.Write indentation, even on first call, is correct.
+            Debug.Indent();
+            VerifyLogged(() => Debug.Write("pizza"),        new string(' ', Debug.IndentLevel * Debug.IndentSize) +  "pizza");
+            Debug.Unindent();
+
+            Debug.WriteLine(""); // neutralize indentation for next WriteLine call: refer to nameof(WriteLine_WontIndentAfterWrite)
+        }
+
+        [Fact]
+        public void Debug_WriteLine_Indents()
+        {
+            // This test when run alone verifies Debug.WriteLine indentation, even on first call, is correct.
+            Debug.Indent();
+            VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', Debug.IndentLevel * Debug.IndentSize) +  "pizza" + Environment.NewLine);
+            Debug.Unindent();
+        }
+
+        [Fact]
+        public void Trace_Write_Indents()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            // This test when run alone verifies Trace.Write indentation, even on first call, is correct.
+            Trace.Indent();
+            VerifyLogged(() => Trace.Write("pizza"),        new string(' ', Trace.IndentLevel * Trace.IndentSize) +  "pizza");
+            Trace.Unindent();
+            
+            Trace.WriteLine(""); // neutralize indentation for next WriteLine call: refer to nameof(WriteLine_WontIndentAfterWrite)
+        }
+
+        [Fact]
+        public void Trace_WriteLine_Indents()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            // This test when run alone verifies Debug.WriteLine indentation, even on first call, is correct.
+            Debug.Indent();
+            VerifyLogged(() => Trace.WriteLine("pizza"),    new string(' ', Trace.IndentLevel * Trace.IndentSize) +  "pizza" + Environment.NewLine);
+            Debug.Unindent();
+        }
+
+        [Fact]
+        public void Debug_WriteLine_WontIndentAfterWrite()
+        {            
             Debug.Indent();
             int expectedIndentation = Debug.IndentLevel * Debug.IndentSize;
             
             VerifyLogged(() => Debug.Write("pizza"),        new string(' ', expectedIndentation) +  "pizza");
             
+            // WriteLine wont indent after Write:
             VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', 0) +                    "pizza" + Environment.NewLine);
 
             VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', expectedIndentation) +  "pizza" + Environment.NewLine);
             VerifyLogged(() => Debug.Write("pizza"),        new string(' ', expectedIndentation) +  "pizza");
+
+            // WriteLine wont indent after Write:
             VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', 0) +                    "pizza" + Environment.NewLine); 
             VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', expectedIndentation) +  "pizza" + Environment.NewLine);
             Debug.Unindent();
         }
 
         [Fact]
-        public void WriteLine_Indents()
+        public void Trace_WriteLine_WontIndentAfterWrite()
         {
-            // This test when run alone verifies Debug.WriteLine indentation, even on first call, is correct.
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            Trace.Indent();
+            int expectedIndentation = Trace.IndentLevel * Trace.IndentSize;
+            
+            VerifyLogged(() => Trace.Write("pizza"),        new string(' ', expectedIndentation) +  "pizza");
+            
+            // WriteLine wont indent after Write:
+            VerifyLogged(() => Trace.WriteLine("pizza"),    new string(' ', 0) +                    "pizza" + Environment.NewLine);
+
+            VerifyLogged(() => Trace.WriteLine("pizza"),    new string(' ', expectedIndentation) +  "pizza" + Environment.NewLine);
+            VerifyLogged(() => Trace.Write("pizza"),        new string(' ', expectedIndentation) +  "pizza");
+
+            // WriteLine wont indent after Write:
+            VerifyLogged(() => Trace.WriteLine("pizza"),    new string(' ', 0) +                    "pizza" + Environment.NewLine); 
+            VerifyLogged(() => Trace.WriteLine("pizza"),    new string(' ', expectedIndentation) +  "pizza" + Environment.NewLine);
+            Trace.Unindent();
+        }
+
+        [Fact]
+        public void Debug_WriteLine_SameIndentationForBothTraceAndDebug()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            Trace.Indent();
+            int expected = Debug.IndentSize * Debug.IndentLevel;
+
+            VerifyLogged(() => Debug.WriteLine("pizza"), new string(' ', expected) +  "pizza" + Environment.NewLine);
+            VerifyLogged(() => Trace.WriteLine("pizza"), new string(' ', expected) +  "pizza" + Environment.NewLine);
+
+            // reset
+            Trace.Unindent();
+        }
+
+        [Fact]
+        public void Trace_WriteNull_IndentsEmptyStringProperly()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            Trace.Indent();
+            int expected = Debug.IndentSize * Debug.IndentLevel;
+
+            VerifyLogged(() => Debug.WriteLine(null), new string(' ', expected) + Environment.NewLine);
+            VerifyLogged(() => Trace.WriteLine(null), new string(' ', expected) + Environment.NewLine);
+
+            // reset
+            Trace.Unindent();
+        }
+        
+        [Fact]
+        public void Trace_UpdatingDebugIndentation_UpdatesTraceIndentation_AndViceVersa()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            int before = Debug.IndentSize * Debug.IndentLevel;
+            Assert.Equal(Debug.IndentSize, Trace.IndentSize);
+            Assert.Equal(Debug.IndentSize, Trace.Listeners[0].IndentSize);
+            Assert.Equal(Debug.IndentLevel, Trace.IndentLevel);
+            Assert.Equal(Debug.IndentLevel, Trace.Listeners[0].IndentLevel);
+
+            Debug.IndentLevel = 3;
+            Assert.Equal(3, Trace.IndentLevel);
+            Assert.Equal(3, Debug.IndentLevel);
+            Assert.Equal(3, Trace.Listeners[0].IndentLevel);
+
+            Debug.IndentLevel = 0;
+            Assert.Equal(0, Trace.IndentLevel);
+            Assert.Equal(0, Debug.IndentLevel);
+            Assert.Equal(0, Trace.Listeners[0].IndentLevel);
+
             Debug.Indent();
-            int expectedIndentation = Debug.IndentLevel * Debug.IndentSize;
+            Assert.Equal(1, Trace.IndentLevel);
+            Assert.Equal(1, Debug.IndentLevel);
+            Assert.Equal(1, Trace.Listeners[0].IndentLevel);
 
-            VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', expectedIndentation) +  "pizza" + Environment.NewLine);
+            Trace.Indent();
+            Assert.Equal(2, Trace.IndentLevel);
+            Assert.Equal(2, Debug.IndentLevel);
+            Assert.Equal(2, Trace.Listeners[0].IndentLevel);
 
-            VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', expectedIndentation) +  "pizza" + Environment.NewLine);
-            VerifyLogged(() => Debug.Write("pizza"),        new string(' ', expectedIndentation) +  "pizza");
-            VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', 0) +                    "pizza" + Environment.NewLine); 
-            VerifyLogged(() => Debug.WriteLine("pizza"),    new string(' ', expectedIndentation) +  "pizza" + Environment.NewLine);
             Debug.Unindent();
+            Trace.Unindent();
+            Assert.Equal(0, Trace.IndentLevel);
+            Assert.Equal(0, Debug.IndentLevel);
+            Assert.Equal(0, Trace.Listeners[0].IndentLevel);
+
+            Debug.Unindent();
+            Assert.Equal(0, Trace.IndentLevel);
+            Assert.Equal(0, Debug.IndentLevel);
+            Assert.Equal(0, Trace.Listeners[0].IndentLevel);
+
+            Trace.IndentSize = 7;
+            Assert.Equal(7, Debug.IndentSize);
+            Assert.Equal(7, Trace.Listeners[0].IndentSize);
+
+            Debug.IndentSize = 4;
+            Assert.Equal(4, Trace.IndentSize);
+            Assert.Equal(4, Trace.Listeners[0].IndentSize);
+
+            Debug.IndentLevel = 0; // reset
+            Assert.Equal(before, Debug.IndentSize * Debug.IndentLevel);
+        }
+
+        [Fact]
+        public void Trace_Refresh_ResetsIndentSize()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            int before = Debug.IndentSize * Debug.IndentLevel;
+            Debug.IndentSize = 5;
+            Debug.IndentLevel = 3;
+            Trace.Refresh();
+
+            Assert.Equal(4, Debug.IndentSize);
+            Assert.Equal(3, Debug.IndentLevel);
+                
+            Debug.IndentLevel = 0; // reset
+            Assert.Equal(before, Debug.IndentSize * Debug.IndentLevel);
+        }
+
+        [Fact]
+        public void Trace_ClearTraceListeners_StopsWritingToDebugger()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            VerifyLogged(() => Debug.Write("pizza"), "pizza");
+            VerifyLogged(() => Trace.Write("pizza"), "pizza");
+            Trace.Listeners.Clear();
+            VerifyLogged(() => Debug.Write("pizza"), string.Empty); 
+            VerifyLogged(() => Trace.Write("pizza"), string.Empty);
+            Trace.Refresh();
+            VerifyLogged(() => Debug.Write("pizza"), "pizza"); 
+            VerifyLogged(() => Trace.Write("pizza"), "pizza");
+
+            Debug.WriteLine(""); // neutralize indentation for next WriteLine call: refer to nameof(WriteLine_WontIndentAfterWrite)
         }
 
         [Fact]
@@ -81,7 +262,7 @@ namespace System.Diagnostics.Tests
             string longString = new string('d', 8192);
             VerifyLogged(() => Debug.Write(longString), longString);
 
-            Debug.WriteLine(""); // neutralize next indentation HACK
+            Debug.WriteLine(""); // neutralize indentation for next WriteLine call: refer to nameof(WriteLine_WontIndentAfterWrite)
         }
 
         [Fact]
@@ -90,7 +271,7 @@ namespace System.Diagnostics.Tests
             VerifyLogged(() => Debug.Print("logged"), "logged");
             VerifyLogged(() => Debug.Print("logged {0}", 5), "logged 5");
 
-            Debug.WriteLine(""); // neutralize next indentation HACK
+            Debug.WriteLine(""); // neutralize indentation for next WriteLine call: refer to nameof(WriteLine_WontIndentAfterWrite)
         }
 
         [Fact]
@@ -122,7 +303,7 @@ namespace System.Diagnostics.Tests
             VerifyLogged(() => Debug.WriteIf(true, "logged", "category"), "category: logged");
             VerifyLogged(() => Debug.WriteIf(false, "logged", "category"), "");
 
-            Debug.WriteLine(""); // neutralize next indentation HACK
+            Debug.WriteLine(""); // neutralize indentation for next WriteLine call: refer to nameof(WriteLine_WontIndentAfterWrite)
         }
 
         [Fact]
@@ -139,6 +320,46 @@ namespace System.Diagnostics.Tests
 
             VerifyLogged(() => Debug.WriteLineIf(true, "logged", "category"), "category: logged" + Environment.NewLine);
             VerifyLogged(() => Debug.WriteLineIf(false, "logged", "category"), "");     
+        }
+
+        [Fact]
+        public void TraceWriteIf()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            VerifyLogged(() => Trace.WriteIf(true, 5), "5");
+            VerifyLogged(() => Trace.WriteIf(false, 5), "");
+
+            VerifyLogged(() => Trace.WriteIf(true, 5, "category"), "category: 5");
+            VerifyLogged(() => Trace.WriteIf(false, 5, "category"), "");
+
+            VerifyLogged(() => Trace.WriteIf(true, "logged"), "logged");
+            VerifyLogged(() => Trace.WriteIf(false, "logged"), "");
+
+            VerifyLogged(() => Trace.WriteIf(true, "logged", "category"), "category: logged");
+            VerifyLogged(() => Trace.WriteIf(false, "logged", "category"), "");
+
+            Trace.WriteLine(""); // neutralize indentation for next WriteLine call: refer to nameof(WriteLine_WontIndentAfterWrite)
+        }
+
+        [Fact]
+        public void TraceWriteLineIf()
+        {
+            if (!_testDebugWithTraceListeners) 
+                return;
+
+            VerifyLogged(() => Trace.WriteLineIf(true, 5), "5" + Environment.NewLine);
+            VerifyLogged(() => Trace.WriteLineIf(false, 5), "");
+
+            VerifyLogged(() => Trace.WriteLineIf(true, 5, "category"), "category: 5" + Environment.NewLine);
+            VerifyLogged(() => Trace.WriteLineIf(false, 5, "category"), "");
+
+            VerifyLogged(() => Trace.WriteLineIf(true, "logged"), "logged" + Environment.NewLine);
+            VerifyLogged(() => Trace.WriteLineIf(false, "logged"), "");
+
+            VerifyLogged(() => Trace.WriteLineIf(true, "logged", "category"), "category: logged" + Environment.NewLine);
+            VerifyLogged(() => Trace.WriteLineIf(false, "logged", "category"), "");     
         }
 
         [Theory]
@@ -262,7 +483,14 @@ namespace System.Diagnostics.Tests
         {
             public static readonly WriteLogger s_instance = new WriteLogger();
 
-            private WriteLogger() { }
+            private WriteLogger()
+            {
+                if (DebugTests._testDebugWithTraceListeners)
+                {
+                    // Wires up TraceListeners to Debug
+                    Assert.Equal(1, Trace.Listeners.Count);
+                }
+            }
 
             public string LoggedOutput { get; private set; }
