+++ src/System.Diagnostics.Debug/tests/System.Diagnostics.Debug.Tests.csproj
@@ -21,7 +21,6 @@
     <Compile Include="DebuggerTests.cs" Condition="'$(TargetGroup)' == 'netcoreapp'" />
     <Compile Include="DebuggerVisualizerAttributeTests.cs" />
     <Compile Include="DebugTests.cs"  Condition="'$(TargetGroup)' == 'netcoreapp'"/>
-    <Compile Include="TraceDebugTests.cs"  Condition="'$(TargetGroup)' == 'netcoreapp'"/>
     <Compile Include="EmptyAttributeTests.cs" />
     <Compile Include="XunitAssemblyAttributes.cs" />
   </ItemGroup>
